\name{as.bayestwin}
\alias{as.bayestwin}
\alias{as.Bayestwin}
\title{Coerce object to bayestwin object}

\usage{as.bayestwin(x,\dots)}

\arguments{
    \item{x}{an object}
    \item{\dots}{unused arguments for compatibility with generic \code{class}}
}

\description{
    the \code{as.bayestwin} function coerces a general R object to a bayestwin object.}

\author{Inga Schwabe}
\keyword{utilities}
