\name{results}
\docType{data}
\alias{results}
\title{Sample output of a BayesTwin analysis}
\description{ Sample MCMC output from \code{IRTtwin} of an analysis of simulated data.The data were simulated and analysed using an ACE and 1 PL model including genotype-environment interaction (GxE).
}
\usage{data(results)}
\format{An \code{bayestwin} object}
\source{
  Schwabe, I. (submitted for publication). BayesTwin - An R Package for Bayesian Inference of Item-Level Hierarchical Twin Data. 
}
\keyword{datasets}
