% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/simulate_twin_data.R
\name{simulate_twin_data}
\alias{simulate_twin_data}
\title{Simulation of twin data}
\usage{
simulate_twin_data(nmz = 300, ndz = 500, var_a = 0.5, var_c = 0.3,
  var_e = 0.2, model = "ACE", n_items = 0)
}
\arguments{
\item{nmz}{refers tot the total number of MZ twins. Defaults to 300}

\item{ndz}{refers tot the total number of DZ twins. Defaults to 500}

\item{var_a}{refers to variance due to additive genetic effects. Defaults to 0.5}

\item{var_c}{refers to variance due to shared-environmental effects. Defaults to 0.3}

\item{var_e}{refers to variance due to unique-environmental effects. Defautls to 0.2}

\item{model}{refers to used model ("ACE", "ADE"). Defaults to "ACE"}

\item{n_item}{refers to the number of items. Defaults to 0 (sumscores are simulated)}
}
\description{
This function simulates twin data. The data can be generated as sumscores but also on item level. Item data is simulated under
a 1 PL Rasch model.
}
\examples{
simulate_twin_data(n_items = 60)
Simualtes twin data on 60 items with the other parameters set to their default values.
}
\keyword{ACE,}
\keyword{ADE}
\keyword{data,}
\keyword{twin}

